requirements.txt
torch~=1.4.0
numpy~=1.18.4
tensorflow-gpu==1.14.0
scipy~=1.4.1
six~=1.12.0
torchvision~=0.5.0
Keras~=2.3.1
scikit-learn~=0.21.3
Pillow~=7.0.0
requests~=2.22.0
tqdm~=4.46.0
joblib~=0.13.2